refactor: move Round class to its own file for better modularity
