+++ src/server/lib/Tournament.ts
@@ -2,56 +2,17 @@ import {Client, ServerTunnelMessage} from '#types/Clients';
 import Clients from '#lib/Clients';
 import {FastifyInstance} from 'fastify';
 import {Player} from '#lib/Match';
-import PongMatch from '#lib/PongMatch';
 import {RawData} from 'ws';
+import Round from '#lib/Round';
 import SQL from 'sql-template-strings';
 import serializeUserAvatar from '#lib/serializeUserAvatar';
 
-interface Participant extends Player {
+export interface Participant extends Player {
   onSocketMessage?: (data: RawData) => void;
   onSocketClose?: () => void;
 }
 
 const MAX_PARTICIPANTS = 8;
-
-class Round {
-  private server;
-  private rounds: Round[] = [];
-  private participants: Participant[] = [];
-
-  constructor(server: FastifyInstance, size: number) {
-    this.server = server;
-
-    const half = size / 2;
-    if (half > 2)
-      for (let i = 0; i < 2; i++) this.rounds.push(new Round(server, half));
-  }
-
-  public addParticipant(participant: Participant) {
-    this.participants.push(participant);
-  }
-
-  public get firstRounds(): Round[] {
-    if (this.rounds) return this.rounds.map(round => round.firstRounds).flat();
-    return [this];
-  }
-
-  public async start() {
-    for (const round of this.rounds || []) {
-      const result = await round.start();
-      if (result.winner) this.participants.push(result.winner);
-    }
-
-    if (this.participants.length < 2)
-      return {winner: this.participants[0], result: 'empty'};
-
-    const match = new PongMatch(this.server, [
-      this.participants[0],
-      this.participants[1],
-    ]);
-    return await match.start();
-  }
-}
 
 export class Tournament {
   public readonly game = 'tournament';
